api-gateway/requirements.txt
# Common dependencies
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.0

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
